updated support levels
