+++ demisto_sdk/commands/lint/resources/pylint_plugins/partner_level_checker.py
@@ -14,13 +14,6 @@ partner_msg = {
     "W9012": ("return_error should be used in main function. Please add it.",
               "return-error-does-not-exist-in-main",
               "return_error should be used in main function",),
-    "W9016": ("Initialize of params was found outside of main function. Please use demisto.params() only inside main"
-              "func",
-              "init-params-outside-main",
-              "Initialize of params was found outside of main function. Please initialize params only inside main func",),
-    "W9017": ("Initialize of args was found outside of main function. Please use demisto.args() only inside main func",
-              "init-args-outside-main",
-              "Initialize of args was found outside of main function. Please use demisto.args() only inside main func",),
 }
 
 
@@ -36,8 +29,6 @@ class PartnerChecker(BaseChecker):
 
     def visit_call(self, node):
         self._return_error_function_count(node)
-        self._init_params_checker(node)
-        self._init_args_checker(node)
 
     def visit_functiondef(self, node):
         self._try_except_in_main(node)
@@ -96,34 +87,6 @@ class PartnerChecker(BaseChecker):
 
         except TypeError:
             yield node
-
-    def _init_params_checker(self, node):
-        try:
-            if node.func.attrname == 'params' and node.func.expr.name == 'demisto':
-                check_param = True
-                parent = node.parent
-                while check_param and parent:
-                    if isinstance(parent, astroid.FunctionDef) and parent.name == 'main':
-                        check_param = False
-                    parent = parent.parent
-                if check_param:
-                    self.add_message("init-params-outside-main", node=node)
-        except AttributeError:
-            pass
-
-    def _init_args_checker(self, node):
-        try:
-            if node.func.attrname == 'args' and node.func.expr.name == 'demisto':
-                check_param = True
-                parent = node.parent
-                while check_param and parent:
-                    if isinstance(parent, astroid.FunctionDef) and parent.name == 'main':
-                        check_param = False
-                    parent = parent.parent
-                if check_param:
-                    self.add_message("init-args-outside-main", node=node)
-        except AttributeError:
-            pass
 
 
 def register(linter):
