+++ demisto_sdk/commands/lint/resources/pylint_plugins/xsoar_level_checker.py
@@ -4,7 +4,15 @@ from pylint.interfaces import IAstroidChecker
 xsoar_msg = {
     "W9013": (
         "Docstrings was not found, Please add to function.", "docstring-doesnt-exits",
-        "Ensure to not try except in the main function.",)
+        "Ensure to not try except in the main function.",),
+    "W9014": (
+        "Function arguments are missing type annotations. Please add type annotations",
+        "args-type-annotations-doesnt-exist",
+        "Function arguments are missing type annotations. Please add type annotations",),
+    "W9015": (
+        "Function return value type annotation is missing. Please add type annotations",
+        "return-type-annotations-doesnt-exist",
+        "Function return value type annotation is missing. Please add type annotations",),
 }
 
 
@@ -19,12 +27,23 @@ class XsoarChecker(BaseChecker):
 
     def visit_functiondef(self, node):
         self._docstring_checker(node)
+        self._type_annotations_checker(node)
 
     # -------------------------------------------- Validations--------------------------------------------------
 
     def _docstring_checker(self, node):
         if not node.doc and node.name != 'main':
             self.add_message("docstring-doesnt-exits", node=node)
+
+    def _type_annotations_checker(self, node):
+        annotation = True
+        for ann, args in zip(node.args.annotations, node.args.args):
+            if not ann and args.name != 'self':
+                annotation = False
+        if not annotation and node.name not in ['main', '__init__']:
+            self.add_message("args-type-annotations-doesnt-exist", node=node)
+        if not node.returns and node.name not in ['main', '__init__']:
+            self.add_message("return-type-annotations-doesnt-exist", node=node)
 
 
 def register(linter):
