+++ demisto_sdk/commands/lint/tests/test_pylint_plugin/xsoar_level_checker_test.py
@@ -96,3 +96,186 @@ class TestDocStringChecker(pylint.testutils.CheckerTestCase):
         with self.assertNoMessages():
             self.checker.visit_functiondef(node_a)
             self.checker.visit_functiondef(node_b)
+
+
+class TestTypeAnnotationsChecker(pylint.testutils.CheckerTestCase):
+    """
+    Class which tests that all functions have type annotations .
+    """
+    CHECKER_CLASS = xsoar_level_checker.XsoarChecker
+
+    def test_type_annotations_exists(self):
+        """
+        Given:
+            - String of a code part which is being examined by pylint plugin.
+        When:
+            - return_output exists in the code
+        Then:
+            - Ensure that the correct message id is being added to the message errors of pylint
+        """
+        node_a, node_b = astroid.extract_node("""
+            def test_num1(a: str, b:int) ->str: #@
+                '''
+                function docs
+                '''
+                return "test function"
+            def test_num2(a: bool, b: bool, c:int) -> bool: #@
+                '''
+                function docs
+                '''
+                if a:
+                    return True
+                if b:
+                    return False
+                else:
+                    return None
+        """)
+        assert node_b is not None and node_a is not None
+        with self.assertNoMessages():
+            self.checker.visit_functiondef(node_a)
+            self.checker.visit_functiondef(node_b)
+
+    def test_args_annotations_doesnt_exist(self):
+        """
+        Given:
+            - String of a code part which is being examined by pylint plugin.
+        When:
+            - Two given function, One does not have type annotations and the other does.
+        Then:
+            - Ensure that the correct message id is being added to the message errors of pylint to the relevent function.
+        """
+        node_a, node_b = astroid.extract_node("""
+            def test_num1(a: str, b:int) ->str: #@
+                '''
+                function docs
+                '''
+                return "test function"
+            def test_num2(a, b: bool, c:int) -> bool: #@
+                '''
+                function docs
+                '''
+                if a:
+                    return True
+                if b:
+                    return False
+                else:
+                    return None
+        """)
+        assert node_b is not None and node_a is not None
+        with self.assertAddsMessages(
+                pylint.testutils.Message(
+                    msg_id='args-type-annotations-doesnt-exist',
+                    node=node_b,
+                ),
+        ):
+            self.checker.visit_functiondef(node_a)
+            self.checker.visit_functiondef(node_b)
+
+        node_a, node_b = astroid.extract_node("""
+            def test_num1(a, b) ->str: #@
+                '''
+                function docs
+                '''
+                return "test function"
+            def test_num2(a, b: bool, c:int) -> bool: #@
+                '''
+                function docs
+                '''
+                if a:
+                    return True
+                if b:
+                    return False
+                else:
+                    return None
+        """)
+
+        assert node_b is not None and node_a is not None
+        with self.assertAddsMessages(
+                pylint.testutils.Message(
+                    msg_id='args-type-annotations-doesnt-exist',
+                    node=node_a,
+                ),
+                pylint.testutils.Message(
+                    msg_id='args-type-annotations-doesnt-exist',
+                    node=node_b,
+                ),
+        ):
+            self.checker.visit_functiondef(node_a)
+            self.checker.visit_functiondef(node_b)
+
+    def test_return_type_annotation_doesnt_exist(self):
+        """
+        Given:
+            - String of a code part which is being examined by pylint plugin.
+        When:
+            - Three given function, one function should raise warnings and the others should not
+        Then:
+            - Ensure that the correct message id is being added to the messages of pylint regarding function test_num1
+        """
+        node_a, node_b, node_c = astroid.extract_node("""
+                    def test_num1(a: str, b: str): #@
+                        '''
+                        function docs
+                        '''
+                        return "test function"
+                    def test_num2(a: bool, b: bool, c:int) -> bool: #@
+                        '''
+                        function docs
+                        '''
+                        if a:
+                            return True
+                        if b:
+                            return False
+                        else:
+                            return None
+                    def client(self) -> bool: #@
+                        '''
+                        function docs
+                        '''
+                        return True
+                """)
+
+        assert node_b is not None and node_a is not None
+        with self.assertAddsMessages(
+                pylint.testutils.Message(
+                    msg_id='return-type-annotations-doesnt-exist',
+                    node=node_a,
+                ),
+        ):
+            self.checker.visit_functiondef(node_a)
+            self.checker.visit_functiondef(node_b)
+            self.checker.visit_functiondef(node_c)
+
+    def test_return_type_annotation_exist(self):
+        """
+        Given:
+            - String of a code part which is being examined by pylint plugin.
+        When:
+            - Three given functions,Two of which have return type annotations and one is main which should cause
+             warnings.
+        Then:
+            - Ensure that there are no errors, Check that there is no error message.
+        """
+        node_a, node_b = astroid.extract_node("""
+                    def test_num1(a: str, b: str) -> str: #@
+                        '''
+                        function docs
+                        '''
+                        return "test function"
+                    def test_num2(a: bool, b: bool, c:int) -> bool: #@
+                        '''
+                        function docs
+                        '''
+                        if a:
+                            return True
+                        if b:
+                            return False
+                        else:
+                            return None
+                    def main():
+                        print("main")
+                """)
+        assert node_b is not None and node_a is not None
+        with self.assertNoMessages():
+            self.checker.visit_functiondef(node_a)
+            self.checker.visit_functiondef(node_b)
