+++ demisto_sdk/tests/integration_tests/xsoar_linter_integration_test.py
@@ -18,9 +18,7 @@ files = [
      ['Demisto.log is found, Please remove all demisto.log usage and exchange it with',
       'Main function wasnt found in the file, Please add main()',
       'Do not use return_outputs function. Please return CommandResults object instead.',
-      'Do not use demisto.results function.',
-      'Function arguments are missing type annotations. Please add type annotations',
-      'Function return value type annotation is missing. Please add type annotations']),
+      'Do not use demisto.results function.']),
     (Path(f"{XSOAR_LINTER_PY3_INVALID}"), 3.8, 'certified partner', False, 1,
      ['Sys.exit use is found, Please use return instead.',
       'Sleep is found, Please remove all sleep statements from the code.']),
@@ -43,8 +41,11 @@ files = [
     (Path(f"{XSOAR_LINTER_PY3_INVALID}"), 3.8, 'xsoar', False, 1,
      ['exit is found, Please remove all exit()', 'quit is found, Please remove all quit()']),
     (Path(f"{XSOAR_LINTER_PY3_INVALID_WARNINGS}"), 3.8, 'xsoar', False, 4,
-     ['Docstrings was not found, Please add to function.']),
-    (Path(f"{XSOAR_LINTER_PY3_INVALID_WARNINGS}"), 3.8, 'partner', False, 4,
+     ['Docstrings was not found, Please add to function.',
+      'Function arguments are missing type annotations. Please add type annotations',
+      'Function return value type annotation is missing. Please add type annotations'
+      ]),
+    (Path(f"{XSOAR_LINTER_PY3_INVALID_WARNINGS}"), 3.8, 'certified partner', False, 4,
      ['Initialize of params was found outside of main function. Please use demisto.params() only inside main',
       'Initialize of args was found outside of main function. Please use demisto.args() only inside main func']),
